[WHATSIT-115] Added the function of deleting pod.

+++ util/Extract.py
@@ -6,6 +6,7 @@ import numpy as np
 from moviepy.editor import VideoFileClip
 
 from util import Trans, Comm
+from util.Comm import delete_pod
 
 
 class Extract:
@@ -18,44 +19,51 @@ class Extract:
 
     def execute(self):
         response_data = Trans.request_service('GET', 'http://api.whatsit.net/datasets/' + self.__dataset_id, [])
-        data_set = response_data['data']['data'][0]
-        video_name = data_set['name']
-        source = data_set['source']
 
-        sections = data_set['sections']
-        video_path = Trans.download_file(os.path.join(self.__path, 'temp.mp4')
-                                         , source)
+        if response_data['data']['data'][0]:
+            data_set = response_data['data']['data'][0]
+            video_name = data_set['name']
+            source = data_set['source']
 
-        video = VideoFileClip(filename=video_path, audio=False, verbose=True)
-        zip = zipfile.ZipFile(os.path.join(self.__path, 'temp.zip'), 'w')
-        image_files = []
-        print('[Extracted image file from video]')
-        for i in sections:
-            for k in np.arange(i[0], i[1] + 1, 0.2):
-                image_inform = {'name': str(k) + '.jpg'}
-                image_files.append(image_inform)
-                image_file_path = os.path.join(self.__img_path, str(k) + '.jpg')
-                video.save_frame(filename=image_file_path, t=k)
-                zip.write(image_file_path, os.path.relpath(image_file_path, self.__img_path),
-                          compress_type=zipfile.ZIP_DEFLATED)
-                print(image_file_path)
+            sections = data_set['sections']
+            video_path = Trans.download_file(os.path.join(self.__path, 'temp.mp4')
+                                             , source)
 
-        zip.close()
-        print('\n\nCreated zip file::' + zip.filename)
-        file_url = Trans.upload_file_to_bucket('whatsit-dataset-video', zip.filename,
-                                               key=self.__dataset_id + '/' + video_name + '.zip', is_public=True)
-        print('image_files' + image_files.__str__())
-        print(file_url)
-        print('Deleted temp directory::')
-        shutil.rmtree(self.__path)
-        params = {
-            "type": "video",
-            "data": [{
-                "frames": file_url
-                , "images": image_files
-            }]
-        }
+            video = VideoFileClip(filename=video_path, audio=False, verbose=True)
+            zip = zipfile.ZipFile(os.path.join(self.__path, 'temp.zip'), 'w')
+            image_files = []
+            print('[Extracted image file from video]')
+            for i in sections:
+                for k in np.arange(i[0], i[1] + 1, 0.2):
+                    image_inform = {'name': str(k) + '.jpg'}
+                    image_files.append(image_inform)
+                    image_file_path = os.path.join(self.__img_path, str(k) + '.jpg')
+                    video.save_frame(filename=image_file_path, t=k)
+                    zip.write(image_file_path, os.path.relpath(image_file_path, self.__img_path),
+                              compress_type=zipfile.ZIP_DEFLATED)
+                    print(image_file_path)
 
-        print(params)
-        print('Requested::')
-        print(Trans.request_service('PUT', 'http://api.whatsit.net/datasets/' + self.__dataset_id, params))
+            zip.close()
+            print('\n\nCreated zip file::' + zip.filename)
+            file_url = Trans.upload_file_to_bucket('whatsit-dataset-video', zip.filename,
+                                                   key=self.__dataset_id + '/' + video_name + '.zip', is_public=True)
+            print('image_files' + image_files.__str__())
+            print(file_url)
+            print('Deleted temp directory::')
+            shutil.rmtree(self.__path)
+            params = {
+                "type": "video",
+                "data": [{
+                    "frames": file_url
+                    , "images": image_files
+                }]
+            }
+
+            print(params)
+            print('Requested::')
+            print(Trans.request_service('PUT', 'http://api.whatsit.net/datasets/' + self.__dataset_id, params))
+        else:
+            print('no existing dataset ')
+
+        # Delete pod
+        delete_pod()
